Refactor MetricsChart into its own component
Recharts charts don't appear to be particularly testable, so
whatever.
